Feat: Add nix flake
Using parlant on nixos requires many dependencies that are hard to get right.
This flake has packages for pythonenv with parlant and the tools.

I was not able to use parlant in a project I'm working on and it always got stuck at caching.
With this flake, I can finally use parlant inside a nix environment.

Signed-off-by: Daniel Poelzleithner <git@poelzi.org>